+++ src/main/kotlin/com/jkjamies/cammp/feature/presentationgenerator/presentation/PresentationUiState.kt
@@ -14,6 +14,8 @@ data class PresentationUiState(
 
     val patternMVI: Boolean = true,
     val patternMVVM: Boolean = false,
+    val patternCircuit: Boolean = false,
+    val diMetro: Boolean = false,
     val diHilt: Boolean = true,
     val diKoin: Boolean = false,
     val diKoinAnnotations: Boolean = false,