+++ src/main/kotlin/com/jkjamies/cammp/feature/repositorygenerator/domain/model/RepositoryParams.kt
@@ -3,6 +3,7 @@ package com.jkjamies.cammp.feature.repositorygenerator.domain.model
 import java.nio.file.Path
 
 sealed interface DiStrategy {
+    data object Metro : DiStrategy
     data object Hilt : DiStrategy
     data class Koin(val useAnnotations: Boolean) : DiStrategy
 }